Fixed LCD Display

+++ gh-iot/model/senser/carbonDioxideSensor.h
@@ -5,7 +5,7 @@ typedef struct carbonDioxideSensor* carbonDioxideSensor_t;
 
 carbonDioxideSensor_t carbonDioxideSensor_create(float reportInterval);
 void carbonDioxideSensor_setReportInterval(carbonDioxideSensor_t self,float reportInterval);
-void carbonDioxideSensor_setValue(carbonDioxideSensor_t self,float value);
+void carbonDioxideSensor_setValue(carbonDioxideSensor_t self,int value);
 float carbonDioxideSensor_getReportInterval(carbonDioxideSensor_t self);
-float carbonDioxideSensor_getValue(carbonDioxideSensor_t self);
+int carbonDioxideSensor_getValue(carbonDioxideSensor_t self);
 myTime_t carbonDioxideSensor_getUpdateTime(carbonDioxideSensor_t self);
